Intro commit
The point of this branch is to change the way that normaliztion is done on traces.  If the user wishes, they can specify, for each trace, a "low" region and a "high" region, which will subsequently inform the normalization.  Not all traces must have low and high states.  Mainly, this helps solve the problem of what to do with traces which, after trimming, contain only one state.  In a multi-channel study, these might still be completely valid to include, but the discretization program has no context to assign the state in that case.  This way, the user can specify which parts of a trace they believe to be the "high" state (if present) and "low" state (if present)