Added pre-processing step, "selection" might work
"Selection" is a property, a logical vector which defines which traces are used in the subsetquent analysis.  The default is for all of them to be "1", of course, but this can be changed by viewing the data and deselecting some of them.  This commit is being made without testing the code, so a few things are probably broken at this stage.