First commit
We're going to want the option to search "under the hood", as it were, using regex, for the potential of extra flexibility.  However, this feature is not going to be user-friendly; it will require the user to have a good understanding of both regex and the conventions of the data set.  That said, it might be useful for some.