Added sample filetypes
These files can be used to understand the input file conventions of KERA (and of some of the intended discretization programs).  Files which are ready to input to QuB, HaMMY, ebFRET, and hFRET are included (in case the user would like to get practice analyzing traces with those programs).  Then, in the corresponding folders, the outputs of those analyses are included, and these are able to be imported to KERA for KERA analysis.